% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_duckdb.R
\name{duckModuleScore}
\alias{duckModuleScore}
\title{AddDuckModuleScore}
\usage{
duckModuleScore(
  con,
  features,
  pool = NULL,
  nbin = 24,
  ctrl = 100,
  assay = "RNA",
  name = "Cluster",
  seed = 1,
  search = FALSE,
  slot = "data",
  ...
)
}
\arguments{
\item{con}{duckdb connection object}

\item{features}{A list of vectors of features for expression programs; each
entry should be a vector of feature names}

\item{pool}{List of features to check expression levels against, defaults to
\code{rownames(x = object)}}

\item{nbin}{Number of bins of aggregate expression levels for all
analyzed features}

\item{ctrl}{Number of control features selected from the same bin per
analyzed feature}

\item{assay}{Name of assay to use}

\item{name}{Name for the expression programs; will append a number to the
end for each entry in \code{features} (eg. if \code{features} has three
programs, the results will be stored as \code{name1}, \code{name2},
\code{name3}, respectively)}

\item{seed}{Set a random seed. If NULL, seed is not set.}

\item{search}{Search for symbol synonyms for features in \code{features} that
don't match features in \code{object}? Searches the HGNC's gene names
database; see \code{\link[Seurat]{UpdateSymbolList}} for more details}

\item{slot}{Slot to calculate score values off of. Defaults to data slot (i.e log-normalized counts)}

\item{...}{Extra parameters passed to \code{\link[Seurat]{UpdateSymbolList}}}
}
\value{
A dataframe with a single column of module score, using cells as rownames
}
\description{
A modified version of Seurat::AddModuleScore() function querying expression data
from duckdb backend
}
