% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_duckdb.R
\name{duckRowMeans}
\alias{duckRowMeans}
\title{duckRowMeans}
\usage{
duckRowMeans(
  con,
  assay = "RNA",
  layer = "counts",
  features = NULL,
  cells = NULL,
  featureTable = "featureTbl",
  cellTable = "cellTbl",
  featureColName = "features",
  cellColName = "cells"
)
}
\arguments{
\item{con}{Duckdb connection object}

\item{assay}{Assay name, by default "RNA"}

\item{layer}{Layer name, by default "counts"}

\item{features}{Features to be included to calculate row mean, used to subset raw data}

\item{cells}{Cells to be included to calculate row mean, used to subset raw data}

\item{featureTable}{The table suffix used to query feature table, the real table name is like "RNA__featureTbl"}

\item{cellTable}{The table suffix used to query cell table, the real table name is like "RNA__cellTbl"}

\item{featureColName}{The column name of the featureTbl, by default it's "features"}

\item{cellColName}{The column name of the cellTbl, by default it's "cells"}
}
\value{
A named vector of rowMeans
}
\description{
Calculate the row (feature) mean value of the layer
}
