% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_duckdb.R
\name{listDuckReduction}
\alias{listDuckReduction}
\title{listDuckReduction}
\usage{
listDuckReduction(con)
}
\arguments{
\item{con}{duckdb connection object}
}
\value{
Table names of the reductions stored in the duckdb
}
\description{
Query reduction data from duckdb database
}
