% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_duckdb.R
\name{queryDuckAssays}
\alias{queryDuckAssays}
\title{queryDuckAssays}
\usage{
queryDuckAssays(con)
}
\arguments{
\item{con}{duckdb connection object}
}
\value{
A vector of of the assays, return NULL if none detected.
}
\description{
List assays from a duckdb connection, based on existing of counts and
data tables (e.g. "RNA__conunts", "RNA__data").
}
