% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_duckdb.R
\name{queryDuckCells}
\alias{queryDuckCells}
\title{queryDuckCells}
\usage{
queryDuckCells(con, assay = "RNA", table = "cellTbl", col_name = "cells")
}
\arguments{
\item{con}{duckdb connection object}

\item{assay}{assay name}

\item{table}{the table suffix used to query cell table, the real table name is like "RNA__cellTbl"}

\item{col_name}{column name of the cellTbl, by default it's "cells"}
}
\value{
A vector containing cells
}
\description{
Query cells of a specific assy from duckdb database
}
