% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_duckdb.R
\name{queryDuckFeatures}
\alias{queryDuckFeatures}
\title{queryDuckFeatures}
\usage{
queryDuckFeatures(
  con,
  assay = "RNA",
  table = "featureTbl",
  col_name = "features"
)
}
\arguments{
\item{con}{duckdb connection object}

\item{assay}{assay name}

\item{table}{the table suffix used to query cell table, the real table name is like "RNA__cellTbl"}

\item{col_name}{column name of the cellTbl, by default it's "cells"}
}
\value{
A vector containing features
}
\description{
Query all the features of a specific assay from duckdb database
}
