% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_duckdb.R
\name{queryDuckMeta}
\alias{queryDuckMeta}
\title{queryDuckMeta}
\usage{
queryDuckMeta(con, meta = "metaData", cols = NULL)
}
\arguments{
\item{con}{duckdb connection object}

\item{metaData}{metaData table name, by default metaData}
}
\value{
A list containing gene's expression, each of the element was neamed by gene's name, and 0 value was discarded.
}
\description{
Query meta data from duckdb database
}
