% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_duckdb.R
\name{queryDuckReduction}
\alias{queryDuckReduction}
\title{queryDuckReduction}
\usage{
queryDuckReduction(con, reduction = "umap", nComponents = 2L)
}
\arguments{
\item{con}{duckdb connection object}

\item{reduction}{reduction name to be extract}

\item{nComponents}{number of components (column) to be kept}
}
\value{
A list containing gene's expression, each of the element was neamed by gene's name, and 0 value was discarded.
}
\description{
Query reduction data from duckdb database
}
