% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_duckdb.R
\name{subsetDuckMatrix}
\alias{subsetDuckMatrix}
\title{subsetDuckMatrix}
\usage{
subsetDuckMatrix(
  con,
  assay = "RNA",
  layers = c("counts", "data"),
  features,
  cells
)
}
\arguments{
\item{con}{duckdb connection}

\item{assay}{assay name}

\item{layers}{layers to be subset, by default, both of the counts and data layer will be subset}

\item{features}{selected features}

\item{cells}{selected cells}
}
\description{
Update duckdb matrix (counts/data layer) to retain specific cells and features
}
