% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_duckdb.R
\name{updateDuckMeta}
\alias{updateDuckMeta}
\title{updateDuckMeta}
\usage{
updateDuckMeta(con, assay = "RNA", data)
}
\arguments{
\item{con}{duckdb connection}

\item{data}{meta data (data frame) to be updated, with the first column as 'cell'}
}
\description{
Update the meta data table of a duckdb database
}
