% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_duckdb.R
\name{updateDuckReduction}
\alias{updateDuckReduction}
\title{updateDuckReduction}
\usage{
updateDuckReduction(con, reductions = c("pca", "umap"), data = list())
}
\arguments{
\item{con}{duckdb connection}

\item{reductions}{a vector of the reductions to be udpated}

\item{data}{list of data frames corresponding to the reductions to be updated}
}
\description{
Udpate the reducions data of a duckdb database
}
