% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the scSpotlight App}
\usage{
run_app(
  onStart = set_options,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  dataDir = NULL,
  runningMode = "processing",
  maxSize = 20 * 1000 * 1024^2,
  nCores = 2,
  ...
)
}
\arguments{
\item{onStart}{A function that will be called before the app is actually run.
This is only needed for \code{shinyAppObj}, since in the \code{shinyAppDir}
case, a \code{global.R} file can be used for this purpose.}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. The default value, \code{NULL}, will respect the setting from
any previous calls to  \code{\link[shiny:enableBookmarking]{enableBookmarking()}}. See \code{\link[shiny:enableBookmarking]{enableBookmarking()}}
for more information on bookmarking your app.}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the
request. Note that the entire request path must match the regular
expression in order for the match to be considered successful.}

\item{dataDir}{Direcotry path of the input data, user could put the large dataset in the
direcoty to avoid uploading files}

\item{runningMode}{The running mode of the app. This could be "processing" (default) or "viewer".
The "processing" mode allows users to process input data by Seurat package while in "viewer" mode,
user will only be able to view the dimension reduction result and query gene expressions.}

\item{maxSize}{Maximum allowed total size (in bytes) of global variables identified, see future.globals.maxSize.}

\item{nCores}{Number of the threads to use (by \code{\link[future:plan]{future::plan()}}).}

\item{...}{arguments to pass to golem_opts.
See \code{?golem::get_golem_options} for more details.}
}
\description{
Run the scSpotlight App
}
\examples{
\dontrun{
 ## Run app in processing mode
 run_app()

 ## Run app in viewer mode and load data in dataDir
 run_app(runningMode = "viewer", dataDir = "/path/to/data")

 ## Run app on port 8081, shiny::runApp() options need to be wrapped in a list
 run_app(options = list(port = 8081, host ="0.0.0.0", launch.browser = FALSE), runningMode = "processing")
}

}
