% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSpotlight-package.R
\docType{package}
\name{scSpotlight-package}
\alias{scSpotlight}
\alias{scSpotlight-package}
\title{scSpotlight: An Amazing Shiny App}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: firstname lastname \email{your@email.com}

}
\keyword{internal}
