% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSpotlight-package.R
\docType{package}
\name{scSpotlight-package}
\alias{scSpotlight}
\alias{scSpotlight-package}
\title{scSpotlight: A Single Cell Analysis Shiny App}
\description{
A single cell analysis (viewer) app based on Seurat.
}
\author{
\strong{Maintainer}: Zhixia Xiao \email{obennoname@gmail.com}

}
\keyword{internal}
