% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_duckdb.R
\name{seurat2duckdb}
\alias{seurat2duckdb}
\title{seurat2duckdb}
\usage{
seurat2duckdb(
  object,
  dbFile = "output.duckdb",
  assay = "RNA",
  layers = c("counts", "data"),
  reductions = c("pca", "umap", "tsne"),
  overwrite = TRUE
)
}
\arguments{
\item{object}{seurat object}

\item{dbFile}{output duckdb filename, please ensure no previous db exists, connection will be closed after conversion}

\item{assay}{searat assay used}

\item{layers}{layers to be converted, please ensure at least one of c("counts", "data") exists, scale.data is not supported yet}

\item{reductions}{object reductions to be included, default: c("pca", "umap", "tsne")}

\item{overwrite}{whether to overwrite existing tables, only could be TRUE for now}
}
\description{
Convert seurat object to onDisk duckdb file
}
\examples{
\dontrun{
 pbmc <- LoadSeurat("pbmc3k.Rds")
 seurat2duckdb(object = pbmc, db = "pbmc3k.duckdb")
}

}
