% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_global.R
\name{withSpinner}
\alias{withSpinner}
\title{WithSpinner}
\usage{
withSpinner(
  ui_element,
  type = getOption("spinner.type", default = 3),
  color = getOption("spinner.color", default = "#2c3e50"),
  size = getOption("spinner.size", default = 0.5),
  color.background = getOption("spinner.color.background", "#ffffff"),
  custom.css = FALSE,
  proxy.height = NULL,
  id = NULL,
  image = NULL,
  image.width = NULL,
  image.height = NULL,
  hide.ui = TRUE,
  fill_container = FALSE
)
}
\description{
A modified withSpinner function from shinycssloaders package,
             refer to: https://github.com/daattali/shinycssloaders/pull/77
}
